Undo refactor done on 69765eb, flake8 fixups of previous commits
Moving the heartbeat and uploader code to its own file on 4537ef4
and the subsequent refactor on 69765eb proved to be the wrong way to
remove circular imports from the package, with some deferred imports
in utils.py being the right solution, so this commit undoes that mistake.
