+++ python/lsst/rubintv/production/__init__.py
@@ -26,5 +26,4 @@ from .utils import *
 from .mountTorques import *
 from .monitorPlotting import *
 from .catchupService import *
-from .uploadTools import *
 from .channels import *
